Prometheus celotool command refactor

+++ packages/celotool/src/lib/k8s-cluster/aks.ts
@@ -1,4 +1,5 @@
 import { execCmd, execCmdWithExitOnFailure } from 'src/lib/cmd-utils'
+import { isCelotoolHelmDryRun } from 'src/lib/helm_deploy'
 import { outputIncludes } from '../utils'
 import { BaseClusterConfig, BaseClusterManager, CloudProvider } from './base'
 
@@ -44,14 +45,18 @@ export class AksClusterManager extends BaseClusterManager {
       `aad-pod-identity exists, skipping install`
     )
     if (!aadPodIdentityExists) {
-      console.info('Adding aad-pod-identity helm repository to local helm')
-      await execCmdWithExitOnFailure(
-        `helm repo add aad-pod-identity https://raw.githubusercontent.com/Azure/aad-pod-identity/master/charts`
-      )
-      console.info('Installing aad-pod-identity')
-      await execCmdWithExitOnFailure(
-        `helm install aad-pod-identity aad-pod-identity/aad-pod-identity -n default`
-      )
+      if (isCelotoolHelmDryRun()) {
+        console.info('Skipping aad-pod-identity deployment due to --helmdryrun')
+      } else {
+        console.info('Adding aad-pod-identity helm repository to local helm')
+        await execCmdWithExitOnFailure(
+          `helm repo add aad-pod-identity https://raw.githubusercontent.com/Azure/aad-pod-identity/master/charts`
+        )
+        console.info('Installing aad-pod-identity')
+        await execCmdWithExitOnFailure(
+          `helm install aad-pod-identity aad-pod-identity/aad-pod-identity -n default`
+        )
+      }
     }
   }
